refactor to vec3 package
