+++ src/main/java/sc/iview/AddLine.java
@@ -1,11 +1,9 @@
 package sc.iview;
 
-import cleargl.GLVector;
-import net.imglib2.RealLocalizable;
-import net.imglib2.RealPoint;
 import org.scijava.command.Command;
 import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
+import sc.iview.vec3.DVec3;
 
 @Plugin(type = Command.class, 
 		menuPath = "SciView>Add>Line")
