+++ src/main/java/sc/iview/vec3/DVec3.java
@@ -1,4 +1,4 @@
-package sc.iview;
+package sc.iview.vec3;
 
 import net.imglib2.Localizable;
 import net.imglib2.RealLocalizable;
