+++ src/main/java/sc/iview/vec3/DVec3s.java
@@ -1,4 +1,4 @@
-package sc.iview;
+package sc.iview.vec3;
 
 import cleargl.GLVector;
